[refactor] #98 : detail 폴더 삭제
