--- src/features/mapView/ui/bottomsheets/detail/index.ts
@@ -1,2 +1,0 @@
-export * from "./TransferDetail";
-export * from "./UserCard";
