+++ src/features/mapView/ui/bottomsheets/detailContents/CarDetail.tsx
