+++ src/features/mapView/ui/bottomsheets/detailContents/index.ts
@@ -1,3 +1,4 @@
 export * from "./Path";
 export * from "./FixedButton";
 export * from "./TransferDetail";
+export * from "./CarDetail";
